some major update

+++ resources/views/user/user_list.blade.php
@@ -53,7 +53,7 @@
                                     <td>{{$data->role}}</td>
                                     <td style="text-align: center;">
                                         <a href="/user/edit/{{$data->id}}" class="btn btn-sm btn-primary"><i class="right fas fa-edit"></i></a>
-                                        <a href="/user/destroy/{{$data->id}}" class="btn btn-sm btn-danger"><i class="right fas fa-trash"></i></a>
+                                        <a href="#" data-toggle="modal" data-href="/user/destroy/{{$data->id}}" class="btn btn-sm btn-danger" data-target="#confirm-delete"><i class="right fas fa-trash"></i></a>
                                     </td>
                                 </tr>
                                 @endforeach
@@ -70,5 +70,30 @@
     </div>
     <!-- /.container-fluid -->
 </section>
+
+<div class="modal fade" id="confirm-delete" tabindex="-1" role="dialog" aria-labelledby="myModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <b>Hapus User</b>
+            </div>
+            <div class="modal-body">
+                Hapus user ini? tidak bisa dikembalikan.
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-default" data-dismiss="modal">Batal</button>
+                <a class="btn btn-danger btn-ok"><b>Hapus</b></a>
+            </div>
+        </div>
+    </div>
+</div>
 <!-- /.content -->
+@endsection
+
+@section('js')
+<script>
+    $('#confirm-delete').on('show.bs.modal', function(e) {
+        $(this).find('.btn-ok').attr('href', $(e.relatedTarget).data('href'));
+    });
+</script>
 @endsection
